move CreateMangaList component to "ui" dir
